backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
pandas==2.2.3
numpy==2.0.2
scikit-learn==1.6.0
joblib==1.4.2
google-generativeai==0.8.3
python-dotenv==1.0.1
faker==25.9.2